Move pkg into codecs/av1
Fixes #221

As previously discussed, the pkg directory is not used in Pion projects.
